HACK: libc: don't fail when fallocate is not supported
On some filesystems like ZFS fallocate isn't supported and will not. So
the only workaround without modifying all relevant upper layer users of
fallocate in the Android stack is to silently ignore ENOTSUP and let
everyone think fallocate works as it should. This should not have much
effects and shows to work well after some testing.